+++ src/model/CurrentAccount.java
@@ -2,7 +2,7 @@ package model;
 
 public class CurrentAccount extends BankAccount{
 
-	public CurrentAccount(long balanceAvailable, String numberAccount) {
+	public CurrentAccount(double balanceAvailable, String numberAccount) {
 		super(balanceAvailable, numberAccount);
 	}
 }
